test(wlan): move the GET option wire tests
Fold get_options_test.go into service_test.go, which mirrors the file
declaring the four methods under test. Two edits are not mechanical:
internal/testutil moves under an assert alias so the bare name keeps
meaning pkg/testutil, and a comment claiming those mocks cannot
observe a query is corrected, since they now record RawQuery.

AI-assisted-by: Claude Code
